requirements.txt
flask==3.1.0
gunicorn==23.0.0
pillow>=10.0.0
numpy>=1.24.0
onnxruntime>=1.17.0
rembg>=2.0.50
